consolidate and improve inference
